Fix uptime ratio display
We used to have a monitor called API which was for keyauth.business. At keyauth.business, a slower VPS with apache was proxying requests to main domain. This was something we setup a while ago when we were with FluxCDN and didn't want to pay double to have another domain. Now, it no longer makes sense and today keyauth.business now points to main server as well.